+++ runtime/v2/types.go
@@ -17,7 +17,7 @@ const (
 	FlagHome   = "home"
 )
 
-// ValidateProtoAnnotations validates that the proto annotations are correct.
+// validateProtoAnnotations validates that the proto annotations are correct.
 // More specifically, it verifies:
 // - all services named "Msg" have `(cosmos.msg.v1.service) = true`,
 //