+++ server/v2/cometbft/utils.go
@@ -17,10 +17,10 @@ import (
 	"google.golang.org/protobuf/types/known/anypb"
 
 	v1beta1 "cosmossdk.io/api/cosmos/base/abci/v1beta1"
-	appmanager "cosmossdk.io/core/app"
 	appmodulev2 "cosmossdk.io/core/appmodule/v2"
 	"cosmossdk.io/core/comet"
 	"cosmossdk.io/core/event"
+	"cosmossdk.io/core/server"
 	"cosmossdk.io/core/transaction"
 	errorsmod "cosmossdk.io/errors"
 	consensus "cosmossdk.io/x/consensus/types"
@@ -68,7 +68,7 @@ func splitABCIQueryPath(requestPath string) (path []string) {
 }
 
 func finalizeBlockResponse(
-	in *appmanager.BlockResponse,
+	in *server.BlockResponse,
 	cp *cmtproto.ConsensusParams,
 	appHash []byte,
 	indexSet map[string]struct{},
@@ -99,7 +99,7 @@ func intoABCIValidatorUpdates(updates []appmodulev2.ValidatorUpdate) []abci.Vali
 	return valsetUpdates
 }
 
-func intoABCITxResults(results []appmanager.TxResult, indexSet map[string]struct{}) []*abci.ExecTxResult {
+func intoABCITxResults(results []server.TxResult, indexSet map[string]struct{}) []*abci.ExecTxResult {
 	res := make([]*abci.ExecTxResult, len(results))
 	for i := range results {
 		if results[i].Error != nil {
@@ -146,7 +146,7 @@ func intoABCIEvents(events []event.Event, indexSet map[string]struct{}) []abci.E
 	return abciEvents
 }
 
-func intoABCISimulationResponse(txRes appmanager.TxResult, indexSet map[string]struct{}) ([]byte, error) {
+func intoABCISimulationResponse(txRes server.TxResult, indexSet map[string]struct{}) ([]byte, error) {
 	indexAll := len(indexSet) == 0
 	abciEvents := make([]*abciv1.Event, len(txRes.Events))
 	for i, e := range txRes.Events {