+++ simapp/v2/app_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
 
-	app2 "cosmossdk.io/core/app"
 	"cosmossdk.io/core/comet"
 	context2 "cosmossdk.io/core/context"
+	"cosmossdk.io/core/server"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/core/transaction"
 	"cosmossdk.io/log"
@@ -53,8 +53,10 @@ func NewTestApp(t *testing.T) (*SimApp[transaction.Tx], context.Context) {
 	// generate genesis account
 	senderPrivKey := secp256k1.GenPrivKey()
 	acc := authtypes.NewBaseAccount(senderPrivKey.PubKey().Address().Bytes(), senderPrivKey.PubKey(), 0, 0)
+	accAddr, err := app.txConfig.SigningContext().AddressCodec().BytesToString(acc.GetAddress())
+	require.NoError(t, err)
 	balance := banktypes.Balance{
-		Address: acc.GetAddress().String(),
+		Address: accAddr,
 		Coins:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdkmath.NewInt(100000000000000))),
 	}
 
@@ -80,7 +82,7 @@ func NewTestApp(t *testing.T) (*SimApp[transaction.Tx], context.Context) {
 
 	_, newState, err := app.InitGenesis(
 		ctx,
-		&app2.BlockRequest[transaction.Tx]{
+		&server.BlockRequest[transaction.Tx]{
 			Time:      time.Now(),
 			Hash:      bz[:],
 			ChainId:   "theChain",
@@ -123,7 +125,7 @@ func MoveNextBlock(t *testing.T, app *SimApp[transaction.Tx], ctx context.Contex
 
 	_, newState, err := app.DeliverBlock(
 		ctx,
-		&app2.BlockRequest[transaction.Tx]{
+		&server.BlockRequest[transaction.Tx]{
 			Height:  height + 1,
 			Time:    time.Now(),
 			Hash:    bz[:],