+++ x/auth/client/cli/broadcast.go
@@ -17,7 +17,7 @@ import (
 // GetBroadcastCommand returns the tx broadcast command.
 func GetBroadcastCommand() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "broadcast [file_path]",
+		Use:   "broadcast <file_path>",
 		Short: "Broadcast transactions generated offline",
 		Long: strings.TrimSpace(`Broadcast transactions created with the --generate-only
 flag and signed with the sign command. Read a transaction from [file_path] and